Allow specification of report ID in setFeatureReport
A new API method signature

int setFeatureReport(byte reportId, byte[] data, int length)

was added.

The previous API was marked as deprecated.  It does not work
consistently on MacOS and Linux/Windows in that the MacOS version
always sends feature report 0, whereas the other two take the first
byte of the caller-supplied buffer as the report ID.